+++ 4/src/main/java/Solution.java
@@ -1,0 +1,35 @@
+import java.io.*;
+import java.math.*;
+import java.security.*;
+import java.text.*;
+import java.util.*;
+import java.util.concurrent.*;
+import java.util.regex.*;
+
+
+public class Solution {
+    private static final Scanner scanner = new Scanner(System.in);
+    //Complete the staircase function below.
+
+    //new fail
+    static void staircase(int n){
+        for (int i = 0; i < n; i++) {
+            for (int j = 0; j < n; j++) {
+                if (j < n - 1 - i) {
+                    System.out.print(" ");
+                } else {
+                    System.out.print("*");
+                }
+            }
+            System.out.println("");
+        }
+    }
+
+    public static void main(String[] args) {
+        int n = scanner.nextInt();
+        scanner.skip("(\r\n|[\n\r\u2028\u2029\u0085])?");
+            staircase(n);
+            scanner.close();
+        }
+}
+
